# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SENTIENT AI API
  description: |
    SENTIENT AI Operating System API
    
    A comprehensive API for AI agent interactions, voice processing,
    channel management, and enterprise features.
  version: 1.0.0
  contact:
    name: SENTIENT Support
    email: support@sentient.ai
    url: https://sentient.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sentient.ai/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Development
...
paths:
  /chat:
    post:
      tags: [Chat]
      summary: Send a chat message
      description: Send a message to the AI agent and receive a response
      operationId: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /chat/stream:
    post:
      tags: [Chat]
      summary: Stream chat response
      description: Send a message and receive streaming response via SSE
      operationId: chatStream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
...
components:
  schemas:
    ChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: User message
        session_id:
          type: string
          description: Optional session ID
        model:
          type: string
          default: gpt-4o
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
        max_tokens:
          type: integer
          minimum: 1
          maximum: 128000
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        response:
          type: string
        session_id:
          type: string
        model:
          type: string
        usage:
          $ref: '#/components/schemas/Usage'
        created_at:
          type: string
          format: date-time
    Usage:
      type: object
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []